fix(contrib): make google_adk_agents MCP toolsets stateless (fixes factory_argument mis-routing)
The previous fix pooled a single McpToolset per provider name in a
worker-process-wide dict shared across all workflow executions, and only
consulted factory_argument the first time a connection opened for that name.
Every later call -- including calls from a different workflow run passing a
different factory_argument -- silently reused that first connection, causing
silent mis-routing for callers that use factory_argument to select a
tenant/credential/backend.

Remove the cross-workflow pool (_ConnectionRecord, _CONNECTIONS,
get_connection, _evict_connection, and the mcp_connection_idle_timeout
parameter) and make TemporalMcpToolSetProvider stateless, mirroring
openai_agents' StatelessMCPServerProvider: each list-tools/call-tool activity
builds a fresh McpToolset via toolset_factory(factory_argument), runs the
operation, and always closes it in a finally block. This fixes the MCP
session/subprocess leak from #1663 while honoring factory_argument on every
call with zero cross-workflow sharing.

Tests rewritten accordingly (test_mcp_pool.py -> test_mcp.py): prove a fresh
toolset per call, close() on every path (success, error, no-match), and that a
later call with a different factory_argument routes with its own argument.

Fixes #1663
